refactor(board): Update default map path to use board-specific structure
