+++ eslint.config.js
@@ -14,5 +14,5 @@ module.exports = {
     'func-names': 'error',
     'prefer-arrow-callback': 'error',
   },
-  ignores: ['**/*.test.js', '**/dist/**', '**/maps/**', '**/rendererMaps/**'],
+  ignores: ['**/*.test.js', '**/dist/**', '**/maps/**'],
 };
