+++ maps/board/default.js
